% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiart.R
\name{asciiart}
\alias{asciiart}
\title{Create asciiart image}
\usage{
asciiart(
  file,
  width = 80,
  text_scaling = 1,
  out_width = 8,
  out_name = NULL,
  text_col = "black",
  chars = c("@", "\%", "#", "*", "+", "=", "-", ":", ".", " ")
)
}
\arguments{
\item{file}{Path to image file or array returned from \code{jpeg::readJPEG(), png::read_PNG()}}

\item{width}{Width of output image in text characters}

\item{text_scaling}{Text scaling factor for tweaking text size (default = 1)}

\item{out_width}{Output image width in inches}

\item{out_name}{File path including name for saved image}

\item{text_col}{Colour of text (default = "black")}

\item{chars}{Character vector set for ascii art (ordered from dark to light)}
}
\description{
Create asciiart image
}
